stanford-cryptography-1/requirements.txt
pycrypto ~= 2.6
gmpy2 ~= 2.0
